backend/pyproject.toml
[project]
name = "myfirstchatbotapp"
version = "0.1.0"
description = ""
authors = [
    {name = "James Reading",email = "james.reading@immersivelabs.com"}
]
readme = "README.md"
requires-python = "^3.13"
dependencies = [
    "fastapi[standard] (>=0.115.14,<0.116.0)",
    "langchain[openai] (>=0.3.26,<0.4.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "langgraph (>=0.5.0,<0.6.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
